pyproject.toml
[tool.poetry]
name = "python-grpc-demo"
version = "0.1.0"
description = ""
authors = ["Aditya Mattos <adityamattos@hey.com>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.51.1"
grpcio-tools = "^1.51.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
run-grpc-server = "python_grpc_demo.server:serve"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
